refactor(e2e): rename import-vault → import-ui + devReset between fixtures
