--- e2e/import-vault.test.js
@@ -1,330 +1,0 @@
-/**
- * Import Vault E2E Test
- *
- * Tests the "I already have a Fast Vault" -> ImportVault flow.
- * Uses a pre-staged test vault file for Detox automation.
- *
- * In __DEV__ mode, the useImportFlow hook checks for a file named
- * `detox-import.vult` (or .bak) in the app's document directory and
- * auto-selects it when the user taps the file picker. This bypasses
- * the system file dialog that Detox cannot automate.
- *
- * NOTE: The staged file is detected, validated, and (if encrypted) the
- * password sheet is shown all in one step. Tests should wait for the
- * password prompt directly after tapping the file picker — the filename
- * badge may not be visible behind the password sheet Modal.
- *
- * Fixture: e2e/fixtures/test-vault.vult (encrypted, password: testpassword123)
- * Generated by: npx tsx e2e/fixtures/generate-test-vault.mts
- */
-const { execSync } = require('child_process')
-const fs = require('fs')
-const path = require('path')
-const FIXTURE_PASSWORD = 'testpassword123'
-const WRONG_PASSWORD = 'wrongpassword999'
-const BUNDLE_ID = 'money.terra.station'
-
-const fixtureVult = path.resolve(
-  __dirname,
-  'fixtures',
-  'test-vault.vult'
-)
-const fixtureBak = path.resolve(
-  __dirname,
-  'fixtures',
-  'test-vault.bak'
-)
-
-/**
- * Stage a file in the simulator's app Documents directory.
- */
-function stageFixture(udid, srcPath, destName) {
-  const appContainer = execSync(
-    `xcrun simctl get_app_container ${udid} ${BUNDLE_ID} data`,
-    { encoding: 'utf8' }
-  ).trim()
-  const documentsDir = path.join(appContainer, 'Documents')
-  execSync(`mkdir -p "${documentsDir}"`)
-  fs.copyFileSync(srcPath, path.join(documentsDir, destName))
-  console.log(`[Import] Staged ${destName} at ${documentsDir}`)
-}
-
-/** Install app, terminate, stage fixture, relaunch. */
-async function setupWithFixture(fixturePath, fixtureName) {
-  const udid = device.id
-
-  await device.launchApp({
-    delete: true,
-    newInstance: true,
-  })
-  await device.terminateApp()
-
-  stageFixture(udid, fixturePath, fixtureName)
-
-  // Relaunch — detoxURLBlacklistRegex only safe on relaunches
-  await device.launchApp({
-    newInstance: true,
-    launchArgs: { detoxURLBlacklistRegex: '.*' },
-  })
-  await device.disableSynchronization()
-}
-
-/**
- * Navigate from AuthMenu (dev mode) through RiveIntro to MigrationHome.
- */
-async function navigateToMigrationHome() {
-  await waitFor(element(by.id('dev-create-fast-vault')))
-    .toBeVisible()
-    .withTimeout(90000)
-  await element(by.id('dev-create-fast-vault')).tap()
-
-  try {
-    await waitFor(element(by.id('enter-vultiverse-cta')))
-      .toBeVisible()
-      .withTimeout(10000)
-    await element(by.id('enter-vultiverse-cta')).tap()
-  } catch {
-    // RiveIntro already transitioned
-  }
-
-  await waitFor(element(by.id('import-vault-button')))
-    .toBeVisible()
-    .withTimeout(90000)
-}
-
-async function navigateToImportScreen() {
-  await waitFor(element(by.id('import-vault-button')))
-    .toBeVisible()
-    .withTimeout(10000)
-  await element(by.id('import-vault-button')).tap()
-  await waitFor(element(by.text('Import your vault share')))
-    .toBeVisible()
-    .withTimeout(30000)
-}
-
-/**
- * Tap file picker and wait for password sheet.
- * The staged file is auto-detected and, if encrypted, the password sheet
- * appears immediately — the filename badge is behind the modal.
- */
-async function pickFileAndWaitForPassword() {
-  await element(by.id('import-file-picker')).tap()
-  await waitFor(element(by.id('decrypt-password-input')))
-    .toBeVisible()
-    .withTimeout(10000)
-}
-
-async function safeEnableSync() {
-  try {
-    await device.enableSynchronization()
-  } catch {}
-}
-
-describe('Import Vault', () => {
-  // ─── Navigation ───────────────────────────────────────────────────
-  describe('Navigation to ImportVault', () => {
-    beforeAll(async () => {
-      await device.launchApp({
-        delete: true,
-        newInstance: true,
-      })
-      await device.disableSynchronization()
-    })
-
-    afterAll(safeEnableSync)
-
-    it('should reach MigrationHome', async () => {
-      await navigateToMigrationHome()
-    })
-
-    it('should navigate to ImportVault screen', async () => {
-      await navigateToImportScreen()
-    })
-
-    it('should show supported file types hint', async () => {
-      await expect(
-        element(by.text('Supported file types: .bak & .vult'))
-      ).toBeVisible()
-    })
-
-    it('should show the Continue button', async () => {
-      await expect(element(by.id('import-continue'))).toExist()
-    })
-
-    it('should navigate back to MigrationHome', async () => {
-      await element(by.id('import-vault-back')).tap()
-      await waitFor(element(by.id('import-vault-button')))
-        .toBeVisible()
-        .withTimeout(5000)
-    })
-  })
-
-  // ─── .vult import with correct password ───────────────────────────
-  describe('Import .vult with correct password', () => {
-    beforeAll(async () => {
-      await setupWithFixture(fixtureVult, 'detox-import.vult')
-    })
-
-    afterAll(safeEnableSync)
-
-    it('should reach ImportVault and trigger file detection', async () => {
-      await navigateToMigrationHome()
-      await navigateToImportScreen()
-      await pickFileAndWaitForPassword()
-    })
-
-    it('should decrypt and import with correct password', async () => {
-      await element(by.id('decrypt-password-input')).typeText(
-        FIXTURE_PASSWORD
-      )
-      await element(by.id('decrypt-continue')).tap()
-
-      await waitFor(element(by.text('You are aboard, Station OG!')))
-        .toBeVisible()
-        .withTimeout(15000)
-    })
-
-    it('should pass vault integrity verification', async () => {
-      await waitFor(element(by.id('verify-all-passed')))
-        .toExist()
-        .withTimeout(15000)
-      await expect(element(by.id('verify-all-passed'))).toHaveText(
-        'all-passed: true'
-      )
-    })
-
-    it('should verify imported vault exists with correct name', async () => {
-      await expect(
-        element(by.id('verify-imported-exists'))
-      ).toHaveText('imported-exists: true')
-      await expect(element(by.id('verify-imported-name'))).toHaveText(
-        'imported-name: true'
-      )
-    })
-
-    it('should verify imported vault has valid keyshare and pubkey', async () => {
-      await expect(
-        element(by.id('verify-imported-has-pubkey'))
-      ).toHaveText('imported-has-pubkey: true')
-      await expect(
-        element(by.id('verify-imported-has-keyshare'))
-      ).toHaveText('imported-has-keyshare: true')
-    })
-
-    it('should complete import and reach main app', async () => {
-      await waitFor(element(by.id('success-back')))
-        .toBeVisible()
-        .withTimeout(5000)
-      await element(by.id('success-back')).tap()
-      await waitFor(element(by.text('Test Import Vault')))
-        .toBeVisible()
-        .withTimeout(10000)
-    })
-  })
-
-  // ─── Persistence ──────────────────────────────────────────────────
-  describe('Persistence — imported vault survives relaunch', () => {
-    beforeAll(async () => {
-      await device.launchApp({
-        newInstance: true,
-        launchArgs: { detoxURLBlacklistRegex: '.*' },
-      })
-      await device.disableSynchronization()
-      await waitFor(element(by.text('Test Import Vault')))
-        .toBeVisible()
-        .withTimeout(30000)
-    })
-
-    afterAll(safeEnableSync)
-
-    it('should not show migration flow on relaunch', async () => {
-      let migrationShown = false
-      try {
-        await waitFor(element(by.id('migration-cta')))
-          .toBeVisible()
-          .withTimeout(5000)
-        migrationShown = true
-      } catch {
-        // Expected — migration is complete
-      }
-
-      if (migrationShown) {
-        throw new Error(
-          'Migration flow should not appear after successful import'
-        )
-      }
-    })
-
-    it('should show imported vault in wallet list', async () => {
-      await expect(element(by.text('Test Import Vault'))).toBeVisible()
-    })
-  })
-
-  // ─── Wrong password ───────────────────────────────────────────────
-  describe('Wrong password shows error', () => {
-    beforeAll(async () => {
-      await setupWithFixture(fixtureVult, 'detox-import.vult')
-    })
-
-    afterAll(safeEnableSync)
-
-    it('should reach password prompt', async () => {
-      await navigateToMigrationHome()
-      await navigateToImportScreen()
-      await pickFileAndWaitForPassword()
-    })
-
-    it('should show error for incorrect password', async () => {
-      await element(by.id('decrypt-password-input')).typeText(
-        WRONG_PASSWORD
-      )
-      await element(by.id('decrypt-continue')).tap()
-
-      await waitFor(element(by.text('Incorrect password, try again')))
-        .toBeVisible()
-        .withTimeout(5000)
-    })
-
-    it('should keep password sheet open after error', async () => {
-      await expect(
-        element(by.id('decrypt-password-input'))
-      ).toBeVisible()
-      await expect(element(by.id('decrypt-continue'))).toBeVisible()
-    })
-  })
-
-  // ─── .bak file extension ─────────────────────────────────────────
-  describe('Import .bak file extension', () => {
-    beforeAll(async () => {
-      await setupWithFixture(fixtureBak, 'detox-import.bak')
-    })
-
-    afterAll(safeEnableSync)
-
-    it('should reach ImportVault and trigger .bak file detection', async () => {
-      await navigateToMigrationHome()
-      await navigateToImportScreen()
-      await pickFileAndWaitForPassword()
-    })
-
-    it('should decrypt and import .bak file successfully', async () => {
-      await element(by.id('decrypt-password-input')).typeText(
-        FIXTURE_PASSWORD
-      )
-      await element(by.id('decrypt-continue')).tap()
-
-      await waitFor(element(by.text('You are aboard, Station OG!')))
-        .toBeVisible()
-        .withTimeout(15000)
-    })
-
-    it('should pass vault integrity verification for .bak import', async () => {
-      await waitFor(element(by.id('verify-all-passed')))
-        .toExist()
-        .withTimeout(15000)
-      await expect(element(by.id('verify-all-passed'))).toHaveText(
-        'all-passed: true'
-      )
-    })
-  })
-})
